Moved generic client file to client folder
